+++ CHANGELOG.md
@@ -9,6 +9,7 @@ BUG FIXES:
 
 * Fix bug that resource `tencentcloud_instance` delete error when instance launch failed.
 * Fix bug that the type of `cluster_type` is wrong in data source `tencentcloud_mongodb_instances`([#242](https://github.com/terraform-providers/terraform-provider-tencentcloud/issues/242)).
+* Fix bug that resource `tencentcloud_eip` unattach error([#233](https://github.com/terraform-providers/terraform-provider-tencentcloud/issues/233)).
 
 ## 1.26.0 (December 09, 2019)